+++ scripts/build.sh
@@ -4,24 +4,33 @@ set -x
 
 desc=$(git describe HEAD)
 mess=$(git show HEAD --no-decorate --pretty=oneline | cut -d ' ' -f2-)
+src=index-bundle-min.js
+
+out=index.html
+if [ -n "$(readlink $out)" ]; then
+    head_ref=$(git symbolic-ref HEAD)
+    branch=${head_ref##*/}
+    out=wip_$branch.html
+fi
 
 case "$mess" in
 Merge\ tag*)
-    desc=$(echo $mess | cut -d"'" -f2)
+    desc=$(echo "$mess" | cut -d"'" -f2)
     base=//github.com/jcorbin/hexant/blob/master
     ;;
 Merge\ branch*)
-    branch=$(echo $mess | cut -d"'" -f2)
+    branch=$(echo "$mess" | cut -d"'" -f2)
     base=//github.com/jcorbin/hexant/blob/$branch
     desc="#$branch $desc"
     ;;
 *)
     base=.
+    src=index-bundle.js
     ;;
 esac
 
 sed \
-    -e "/data-import/s/src=[^>]*>/src=\"index-bundle-min.js\">/" \
+    -e "/data-import/s/src=[^>]*>/src=\"$src\">/" \
     -e "s~BASE~$base~" \
     -e "s/DEV/$desc/" \
     -e "/PROJECT LINK/d" \
@@ -29,7 +38,9 @@ sed \
 
 npm run grammar
 bundle index.js >index-bundle.js
-minify index-bundle.js >index-bundle-min.js
+if [ "$src" == "index-bundle-min.js" ]; then
+    minify index-bundle.js >index-bundle-min.js
+fi
 
-html-inline -i index-tmp.html -o index.html
+html-inline -i index-tmp.html -o "$out"
 rm -f index-tmp.html